[core] move base test case to __init__, add patching helpers
